Ensaio Dialelo para o Tempo Até o Florescimento do Feijoeiro

Dados referentes ao cruzamento dialelo realizado por Arriel et al. (1993), em que foi avaliado o número de dias para o florescimento do feijoeiro para cada progenie.

Format

Um data.frame com 9 observações e 3 variáveis, em que

prog1

Inteiro que indica o progenitor 1 do cruzamento.

prog2

Inteiro que indica o progenitor 2 do cruzamento.

dias

Número médio de dias para o florescimento do feijoeiro.

Source

RAMALHO et al. (2005), Exemplo 12.10, pág. 227.

Arriel, E., Ramalho, M. A. P., Pacheco, C. A. P. Expected and Realized Gains in the CMS-39 Maize Population after three Cycles of Selection. Revista Brasileira de Genética. Ribeirão Preto. v.16, n.4, p.1013. dez. 1993

Examples

str(RamalhoEg12.10)

#> 'data.frame':	9 obs. of  3 variables:
#>  $ prog1: int  1 1 1 1 2 2 3 3 4
#>  $ prog2: int  2 3 4 5 3 4 4 5 5
#>  $ dias : num  48 31 28.7 31.3 30.7 26 29.7 26 28

RamalhoEg12.10

#>   prog1 prog2 dias
#> 1     1     2 48.0
#> 2     1     3 31.0
#> 3     1     4 28.7
#> 4     1     5 31.3
#> 5     2     3 30.7
#> 6     2     4 26.0
#> 7     3     4 29.7
#> 8     3     5 26.0
#> 9     4     5 28.0

# Número de genitores.
n <- with(RamalhoEg12.10, unique(c(prog1, prog2)))
RamalhoEg12.10 <- transform(RamalhoEg12.10,
                            prog1 = factor(prog1, levels = n),
                            prog2 = factor(prog2, levels = n))

# Matriz do cruzamento dialelo.
X1 <- model.matrix(~0 + prog1, data = RamalhoEg12.10)
X2 <- model.matrix(~0 + prog2, data = RamalhoEg12.10)
X <- 0.5 * (X1 + X2)